+++ CHANGELOG.md
@@ -1,16 +1,19 @@
 # CHANGELOG
 
-## v0.6.1
+## v0.7.0
 
 - Parser
   - Add `debug` option to `AdvancedBuilder.expand`.
   - Auto resolve R-S conflicts when `AdvancedBuilder.expand`.
   - Allow `'*'` as the `next` when resolve RS/RR conflicts.
   - Add `ParserContext.gr/firstSets/followSets`.
+  - **Breaking Change**: `ParserBuilder.generateResolvers/checkConflicts` is now private, and `ParserBuilder.checkAll` is removed. Use `options` in `ParserBuilder.build` instead.
 - Fix
   - Parser can detect R-S conflict like: `` exp: `a b | a b c` `` where `b` is a terminator.
   - Parser won't print `[user resolved RR]` when it's actually an R-S conflict.
 - Optimize code & comments.
+- Optimize performance.
+  - Merge `ParserBuilder.generateResolvers/checkConflicts` to `ParserBuilder.build` to avoid unnecessary DFA build.
 - More tests.
 
 ## v0.6.0